refactor: optimize search performance and improve code organization
- Implement O(1) search using pre-built substring indexes for names, keywords, and aliases
- Add comprehensive JSDoc for isValidEmoji explaining regex ranges and limitations
- Split monolithic index.ts into cleaner modules with backward-compat.ts
- Maintain 100% backward compatibility while improving maintainability

The search function now uses pre-computed indexes instead of nested loops,
significantly improving performance for large emoji datasets.
